Delete button now works, checks for enrolments. Working in courses, have some issues with deleting lecturers

+++ src/components/DeleteBtn.js
@@ -4,28 +4,55 @@ import { useState } from "react";
 
 
 //creates and exports
-const DeleteBtn = () => {
+const DeleteBtn = ({id, resource, anotherResource, deleteCallBack, data}) => {
     let [isLoading, setIsLoading] = useState(false);
     //const navigate = useNavigate();
 
+    // const [course, setCourse] = useState([]);
+
     //pass in an id as a prop
-    const onDelete = (id, resource, deleteCallBack) => {
+    const onDelete = () => {
+
         isLoading = true;
         let token = localStorage.getItem('token');
 
-        axios.delete(`https://college-api.vercel.app/api/${resource}/${id}`, {
-            headers: {
-            Authorization: `Bearer ${token}`
-            }
-        })
-            .then(response => {
-                console.log(response.data);
-                deleteCallBack(id);
-                //navigate('/festivals');
+        console.log(id);
+        console.log(data[anotherResource]);
+            
+        let listOfDeleteRequests = data[anotherResource].map((current, index)=>
+        axios.delete(`https://college-api.vercel.app/api/${anotherResource}/${current.id}`, 
+            {headers: { Authorization: `Bearer ${token}` }}
+        ));
+            // log the contents of listOfDeleteRequests
+            axios.all(listOfDeleteRequests) //this line deletes all enrolments first and then runs the then() method below
+                .then((response) => {
+
+                    axios.delete(`https://college-api.vercel.app/api/${resource}/${id}`, {
+                    headers: { Authorization: `Bearer ${token}` }
             })
-            .catch(err => {
-                console.error(err.response.data);
+                .then((response) => {
+                 //do something
+                 if (data[anotherResource].length <= 0) {
+                    axios.delete(`https://college-api.vercel.app/api/${resource}/${id}`, {
+                        headers: {
+                        Authorization: `Bearer ${token}`
+                        }
+                    })
+                        .then(response => {
+                            console.log(response.data);
+                            deleteCallBack(id);
+                            //navigate('/festivals');
+                        })
+                        .catch(err => {
+                            console.error(err.response.data);
+                        })
+                }
             })
+                .catch((error) => {
+                    console.log(error);
+            });
+
+        });
     }
     return (
         <>
@@ -37,3 +64,4 @@ const DeleteBtn = () => {
 }
 
 export default DeleteBtn;
+
